+++ src/components/Paper/PaperContainer.jsx
@@ -1,21 +1,14 @@
 import { useState } from 'react';
-import leftArrow from '../../assetes/images/arrow-left.svg';
-import rightArrow from '../../assetes/images/arrow-right.svg';
-import { SLIDE_CONTAINER_SIZE } from '../../constants/paperConstants';
 import useFetchData from '../../hooks/useFetchData';
-import getCarouselIndex from '../../utils/getCarouselIndex';
+import LeftArrowButton from './Button/LeftArrowButton';
+import RightArrowButton from './Button/RightArrowButton';
 import Paper from './Paper';
 import * as S from './PaperContainer.style';
 
 export default function PaperContainer() {
-  const [carouselIndex, setCarouselIndex] = useState(0);
-
   const paperData = useFetchData('recipients/');
-
-  const lastIndexOfCarousel = getCarouselIndex(paperData);
-
-  const isVisibleRightArrow = carouselIndex > lastIndexOfCarousel;
-  const isVisibleLeftArrow = carouselIndex !== 0;
+  const paperLength = paperData ? paperData.results.length : 0;
+  const [carouselIndex, setCarouselIndex] = useState(0);
 
   const handleArrowClick = increment => {
     setCarouselIndex(prev => prev + increment);
@@ -23,11 +16,10 @@ export default function PaperContainer() {
 
   return (
     <S.Wrapper>
-      {isVisibleLeftArrow && (
-        <S.LeftArrowBox onClick={() => handleArrowClick(SLIDE_CONTAINER_SIZE)}>
-          <S.Arrow src={leftArrow} alt="left-arrow" />
-        </S.LeftArrowBox>
-      )}
+      <LeftArrowButton
+        carouselIndex={carouselIndex}
+        handleArrowClick={handleArrowClick}
+      />
       <S.SlideContainer>
         <S.SlideElement $carouselIndex={carouselIndex}>
           {paperData &&
@@ -36,13 +28,11 @@ export default function PaperContainer() {
             ))}
         </S.SlideElement>
       </S.SlideContainer>
-      {isVisibleRightArrow && (
-        <S.RightArrowBox
-          onClick={() => handleArrowClick(-SLIDE_CONTAINER_SIZE)}
-        >
-          <S.Arrow src={rightArrow} alt="right-arrow" />
-        </S.RightArrowBox>
-      )}
+      <RightArrowButton
+        carouselIndex={carouselIndex}
+        paperLength={paperLength}
+        handleArrowClick={handleArrowClick}
+      />
     </S.Wrapper>
   );
 }
