+++ src/utils/getCarouselIndex.js
@@ -3,8 +3,8 @@ import {
   SLIDE_CONTAINER_SIZE
 } from '../constants/paperConstants';
 
-export default function getCarouselIndex(paperData) {
-  const totalItems = paperData ? paperData.results.length : 0;
+export default function getCarouselIndex(paperLength) {
+  const totalItems = paperLength;
   const totalPages = Math.ceil(totalItems / ITEMS_PER_CONTAINER) - 1;
 
   const lastIndexOfCarousel = -(totalPages * SLIDE_CONTAINER_SIZE);
